updates to the Debian package build script

+++ tools/create-debian-packages.sh
@@ -1,9 +1,14 @@
 #!/bin/bash
 
-if [ `id -u` != 0 ]; then
-	echo "This script must be run as root."
-	exit 2
-fi
+DEST_DIR_ROOT=$HOME/packages
+CURR_DIR=`pwd`
+BZR_IGRAPH_ROOT=$HOME/bzr/igraph
+
+SIGN=yes
+SIGNING_GPG_KEY=D6360653
+
+#####################################################################
+
 
 if [ $# -ne 3 ]; then
     echo "Usage: $0 igraph_version debian_revision series"
@@ -21,21 +26,32 @@ IGRAPH_VERSION=$1
 DEBIAN_REVISION=$2
 SERIES=$3
 
-DEST_DIR=$HOME/packages/$SERIES
-CURR_DIR=`pwd`
-BZR_IGRAPH_ROOT=$HOME/bzr/igraph
+DEST_DIR=${DEST_DIR_ROOT}/${SERIES}
+
+if [ x$SIGN == xyes ]; then
+  SIGN_OPTIONS="-k0x${SIGNING_GPG_KEY}"
+else
+  SIGN_OPTIONS="-us -uc"
+fi
+
+BAZAAR_BRANCH=`echo $IGRAPH_VERSION | awk 'BEGIN { FS="."; OFS="." } ; { printf "%s.%s-main", $1, $2 }'`
 
 function prechecks {
   if [ ! -d ${BZR_IGRAPH_ROOT} ]; then
     echo ${BZR_IGRAPH_ROOT} does not exist or is not a directory!
     exit 1
   fi
+
+  if [ x$SIGN == xyes ]; then
+    which gpg >/dev/null || ( echo "gpg not installed, exiting"; exit 1 )
+    gpg --list-secret-keys ${SIGNING_GPG_KEY} >/dev/null || ( echo "Secret key not found: ${SIGNING_GPG_KEY}, exiting"; exit 1 )
+  fi
 }
 
 function install_build_dependencies {
   apt-get update
   apt-get -y --no-install-recommends install \
-          debhelper devscripts fakeroot automake autoconf gcc g++ \
+          cdbs debhelper devscripts fakeroot automake autoconf gcc g++ \
           pkg-config flex bison build-essential libxml2-dev libglpk-dev \
           libarpack2-dev libgmp3-dev libxml2-dev libblas-dev liblapack-dev \
           python-all-dev python-central python-epydoc texlive-latex-base 
@@ -46,7 +62,7 @@ function make_destdir {
 }
 
 function bazaar_update {
-  pushd ${BZR_IGRAPH_ROOT}/0.6-main
+  pushd ${BZR_IGRAPH_ROOT}/${BAZAAR_BRANCH}
   bzr update
   popd
 }
@@ -59,20 +75,26 @@ function create_igraph_debian_pkg {
   fi
   tar -xvvzf igraph_$1.orig.tar.gz
   cd igraph-$1
-  cp -r ${BZR_IGRAPH_ROOT}/0.6-main/debian .
-  cat debian/changelog.in | sed -e "s/@VERSION@/@VERSION@-${DEBIAN_REVISION}/g" >debian/changelog.in.new
-  mv debian/changelog.in.new debian/changelog.in
+  cp -r ${BZR_IGRAPH_ROOT}/${BAZAAR_BRANCH}/debian .
+  if [ x$2 != x0 ]; then
+    cat debian/changelog.in | sed -e "s/@VERSION@/@VERSION@-$2/g" >debian/changelog.in.new
+    mv debian/changelog.in.new debian/changelog.in
+  fi
   debian/prepare
   cat debian/changelog | sed -e "s/unstable/${SERIES}/g" >debian/changelog.new
   mv debian/changelog.new debian/changelog
-  # dpkg-buildpackage -tc -rfakeroot
-  debuild -b -us -uc
-  debuild -S -sa
+  debuild -b ${SIGN_OPTIONS}
+  debuild -S -sa ${SIGN_OPTIONS}
   cd ..
 }
 
 function install_igraph_debian_pkg {
-  dpkg -i libigraph0_$1-$2_*.deb libigraph-dev_$1-$2_*.deb || exit 3
+  if [ x$2 != x0 ]; then
+    FULLVERSION=$1-$2
+  else
+    FULLVERSION=$1
+  fi
+  ${SUDO} dpkg -i libigraph0_${FULLVERSION}_*.deb libigraph-dev_${FULLVERSION}_*.deb || exit 3
 }
 
 function remove_igraph_debian_pkg {
@@ -80,16 +102,22 @@ function remove_igraph_debian_pkg {
 }
 
 function compile_python_interface {
-  wget -O python-igraph_$1.orig.tar.gz http://pypi.python.org/packages/source/p/python-igraph/python-igraph-$1.tar.gz
+  if [ -f ${CURR_DIR}/python-igraph_$1.orig.tar.gz ]; then
+    cp ${CURR_DIR}/python-igraph_$1.orig.tar.gz .
+  else
+    wget -O python-igraph_$1.orig.tar.gz http://pypi.python.org/packages/source/p/python-igraph/python-igraph-$1.tar.gz
+  fi
   tar -xvvzf python-igraph_$1.orig.tar.gz
   cd python-igraph-$1
-  cat debian/changelog.in | sed -e "s/@VERSION@/@VERSION@-${DEBIAN_REVISION}/g" >debian/changelog.in.new
+  if [ x$2 != x0 ]; then
+    cat debian/changelog.in | sed -e "s/@VERSION@/@VERSION@-$2/g" >debian/changelog.in.new
+  fi
   mv debian/changelog.in.new debian/changelog.in
   debian/prepare
   cat debian/changelog | sed -e "s/unstable/${SERIES}/g" >debian/changelog.new
   mv debian/changelog.new debian/changelog
-  debuild -b -us -uc
-  # debuild -S -sa
+  debuild -b ${SIGN_OPTIONS}
+  debuild -S -sa ${SIGN_OPTIONS}
   cd ..
 }
 
@@ -102,18 +130,24 @@ function remove_build_dir {
   rm -rf ${BUILD_DIR}
 }
 
-prechecks
+prechecks || exit 1
 
 BUILD_DIR=`mktemp -d`
 cd ${BUILD_DIR}
 trap "{ cd /; rm -rf \"${BUILD_DIR}\"; exit 255; }" INT EXIT TERM
 
-install_build_dependencies
+if [ `id -u` != 0 ]; then
+  echo "Not running as root, skipping installation of build-deps."
+  SUDO=sudo
+else
+  install_build_dependencies
+  SUDO=
+fi
 make_destdir
 bazaar_update
-create_igraph_debian_pkg ${IGRAPH_VERSION}
+create_igraph_debian_pkg ${IGRAPH_VERSION} ${DEBIAN_REVISION}
 install_igraph_debian_pkg ${IGRAPH_VERSION} ${DEBIAN_REVISION}
-compile_python_interface ${IGRAPH_VERSION}
+compile_python_interface ${IGRAPH_VERSION} ${DEBIAN_REVISION}
 remove_igraph_debian_pkg
 move_packages
 remove_build_dir
